Support single row list with crawl.
`$isMultiColumnList` of `BEAR\Resource\Linker::isList()` doens't become `true` unless the list contains more than one line, so the list with only one line wasn't evaluated as a list.
`$isMultiColumnList` has been changed to `$isMultiColumnMultiRowList` and `$isMultiColumnList` is set to `true` if `$value` is an array containing an array.

BTW, I'm not sure whether need be left `$isMultiColumnList` as `$isMultiColumnMultiRowList` so I left it as `$isMultiColumnMultiRowList`.

- Previous Behavior

Could't be `true`

```
$value = [
	[
		'id' => 1,
		'name' => 'hoge'
	]
];
```

Could be `true`

```
$value = [
	[
		'id' => 1,
		'name' => 'hoge'
	],
	[
		'id' => 2,
		'name' => 'fuga'
	]
];
```

- Behavior by this commit

Could be `true`

```
$value = [
	[
		'id' => 1,
		'name' => 'hoge'
	]
];
```

```
$value = [
	[
		'id' => 1,
		'name' => 'hoge'
	],
	[
		'id' => 2,
		'name' => 'fuga'
	]
];
```