Implement HttpRequestInterface and update tests
The HttpRequestCurl now implements a new HttpRequestInterface, resulting in a more robust and flexible design. The switch over from HttpRequestCurl to HttpRequestInterface in HttpResourceObject and HttpClientModule is reflected aswell.  Additionlly, tests and server-side scripts have been updated to handle HTML responses and accompanying headers properly. This enhances the system's capability to handle different types of requests and makes the test suite more comprehensive.